requirements.txt
streamlit
Pillow
PyPDF2
python-docx
pdf2image
google-cloud-vision
langdetect
wordninja
yake
spacy
pytextrank
